pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
dependencies = [
    "accelerate>=1.9.0",
    "chatterbox-tts @ git+https://github.com/anthonywu/chatterbox.git@dev",
    "elevenlabs[pyaudio]>=1.0,<2.0",
    "platformdirs>=4.3,<5.0",
    "peft>=0.16.0,<1.0",
    "python-dotenv>=1.1.1",
    "psutil>=7.0,<8.0",
    "torch>=2.6,<2.7",                                                      # chatterbox restriction, find out why later
    "torchaudio>=2.6,<2.7",                                                 # chatterbox restriction, find out why later
    "tqdm>=4.67,<5.0",
]
name = "gensay"
dynamic = ["version"]
description = 'Multi-provider TTS tool compatible with macOS say command'
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = []
authors = [
    { name = "Anthony Wu", email = "pls-file-gh-issue@users.noreply.github.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
]

[project.scripts]
gensay = "gensay.main:main"

[project.urls]
Documentation = "https://github.com/anthonywu/gensay#readme"
Issues = "https://github.com/anthonywu/gensay/issues"
Source = "https://github.com/anthonywu/gensay"

[tool.hatch.version]
path = "src/gensay/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/gensay tests}"

[tool.hatch.envs.test]
extra-dependencies = ["pytest>=8.0.0", "pytest-cov>=5.0.0"]
[tool.hatch.envs.test.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov=gensay --cov-report=term-missing {args:tests}"

[tool.coverage.run]
source_pkgs = ["gensay", "tests"]
branch = true
parallel = true
omit = ["src/gensay/__about__.py"]

[tool.coverage.paths]
gensay = ["src/gensay", "*/gensay/src/gensay"]
tests = ["tests", "*/gensay/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C90", # mccabe complexity
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # unused imports in __init__ files
"tests/*" = ["ARG"]      # unused arguments in tests

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "chatterbox.*",
    "torchaudio.*",
    "torch.*",
    "librosa.*",
    "soundfile.*",
    "pytest.*",
]
ignore_missing_imports = true
